Show 'Opened Windows' in Application panel's frame tree
Show info about windows opened via `window.open()` in the frame tree.
Clicking on the frame tree entry for such a window opens its detail
view on the right. Whether the opened window is able to access the
opener via `window.opener` is displayed in this details view. If the
opener is available, a link to the opener in the elements panel is
generated.

Screenshots:
https://i.imgur.com/ozaI3SG.png
https://i.imgur.com/k2YhtWO.png

Bug: chromium:1107766
Change-Id: Ia2660d6932fcb228dcb204d2e9eadaaf9ca5f00c