[Issues] UMA tracking for the issues panel
This CL add new metrics enums to allow measuring when an issue in the
issue panel is expanded and when a resource is opened.

The corresponding chromium CL is at https://crrev.com/c/2305018

There are two elements that we can't track the telemetry because
they are stopping the propagation of the event for "click" and
"keydown" if the key is the enter key.

I have open two issues to track them separately.

  - Learn More link: https://crbug.com/1108501
  - Source link: https://crbug.com/1108503


Bug: 1100658
Change-Id: I4a6f86f107dc710bc64edbf21670a4a3bd2a3d60