[unittest] Add 'Settings' helper to set up settings storage
This CL cleans up the usage of 'Settings' into a re-usable, general
helper. The helper sets up the local/global settings storage, so
units under test can exercise their integration with the Settings
infrastructure.

This helper is more for integration tests, and tests that utilize
this should be moved to integration tests once they exist.

Please note that tests using the helper should call it in both,
before AND after hooks, to ensure that Settings don't leak between
tests.

R=aerotwist@chromium.org, jacktfranklin@chromium.org

Change-Id: If77ea4b91e3e6e9cc4823d82d030063477f4fac9