[issues] Add detection for third-party SameSite cookie issues
This CL implements 'isCausedByThirdParty' detection for SameSite
cookie issues. Issues we consider caused by a third-party are not
shown by default but can be enabled via a Issues tab setting.

We assume developers have no control over third-party sites,
making reported issues non-actionable.

For SameSite cookies, we compare the cookieUrl with the top-most
frames registered domain. In the future, we want to compare the
cookieUrl with the top-most frames first-party set, which should
be almost a perfect fit.

R=chlily@chromium.org, sigurds@chromium.org

Bug: chromium:1080589
Change-Id: I70ce06aa7c50aee87219b2c24acc2c2a9c806e6c