Grid: Always place grid area labels in the top/left corner of the area
Grid area name labels are supposed to be displayed in the top/left
corner of the corresponding grid areas. And they are, most of the time.

The way this was achieved was by using the minX and minY coordinates
from the area bounds.
However, when the grid container is transformed (e.g. rotated), the minX
and minY might not correspond to the top/left corner anymore.

So now we use the bounds.allPoints[0] data, which always gives us the
x/y coordinates of the top/left corner.

Before: https://imgur.com/aE9Gx3U.jpg
After: https://i.imgur.com/Xi6bNaq.jpg
Bug: 1100384
Change-Id: Icb0acb1f4b1a4d77c0e7df6af18e48bb96608cc7