Fix type reference detection in bridge generator
This CL fixes one particular case when the AST traversal wasn't thorough
enough; if you extend an interface, we need to also check the interface
that's being inherited from for any types that need to be pulled into
the bridge.

For example:

```
interface A {...}
interface B {
  a: A
}
interface C extends B {
  ...
}
```

If the bridge generator decides that interface `C` should be in the
bridge, we need to walk through its parents (in this case, `B`) to check
for any other nested type references that should be included in the
bridge. In this case, because `B` references the `A` interface, we need
to add the `A` interface to the bridge.

Change-Id: Id211e2e02c12fed9ebb7c6456640bf87c3be70ac