Fixes needed to build desktop chrome in an alternate toolchain.
NOT FOR LANDING!

This is a proof-on-concept patch needed as part of the work to
get 'lacros-chrome' to build as part of a build of 'ash-chrome',
using an alternate toolchain.

We do not normally build desktop Chrome in an alternate toolchain,
and so there are various bugs in the build resulting from the fact
that root_build_dir != root_out_dir.

See https://crrev.com/c/2343924 for the main change.

Commit: false

Change-Id: I32fc3620220bc8c20a3c87d926f01e6c88362ca7