eliminate initial timeout in waitForFunction
`waitForFunction` previously always waited 100 ms before checking if the
given function is truthy. We can improve the total e2e test suite running
time by approx. 50 s by eliminating this timeout.  This, unfortunately,
makes some of the tests flake since they relied on this timeout. This CL
attempts to eliminate the timeout and patch the tests that relied on it.
In the cases where a test could not be patched the timeout is inserted in
the test explicitly.

Change-Id: Id92542de85920f95bb922d1b62ddfb63b83a055a