[ts] Type check ui/Action.js with TypeScript
This CL introduces a new class that properly types the various fields
when an 'action' is defined in a 'module.json' file.

Unfortunately, the intersection of JSDoc features supported by both
Closure and TypeScript, doesn't allow this type to be defiend in
pure JSDoc. For this reason a class is introduced that is neverly
actually constructed but serves as a Type definition.

R=aerotwist@chromium.org

Bug: chromium:1011811
Change-Id: I444f23f5ceb224aa33b754c301eb5d393ae74d54