components: Create a PieChart component
Add a web component called PieChart2. The old non-component version
is called PieChart already. We'll replace it with this one once it lands
and then remove the original PieChart and rename this one.

Switch over one user of PieChart (TimelineUIUtils) to use the new
version.

Add tests and examples to the component server.

Change-Id: I563059e0322739d533a95155735fcafbe414e815