[ts] Move Element prototype wrangling to DOMExtension.js
This CL moves 4 overriden methods to DOMExtension.js. These are hard
to type-check with TS, so we don't attempt it and move all prototype
wrangling code into DOMExtension.js.

Widget.__assert is inlined directly, unfortunately, we can't use
console.assert, as some layout test expects a normal Error to be
thrown and not an assertion error.

CL 1 out of * to enable TS for ui/Widget.js.

R=aerotwist@chromium.org

Bug: chromium:1011811
Change-Id: Ibf78bc1aa32f7e81dc3dc303858a7c0f875fdc95