[ts] Store Widget scroll positions in WeakMap instead of the element
This CL changes how the Widget stores and restores scroll positions
of potential child elements. Instead of storing the scroll positions
of the element directly, this CL introduces a new WeakMap that
performs this job.

This solution also has the advantage that TypeScript will also be
happy.

CL 2 out of * to enable TS for ui/Widget.js.

R=jacktfranklin@chromium.org

Bug: chromium:1011811
Change-Id: I3967f49977e19590b015128618f093a250f382f7