[ts] Type-check ui/Widget.js with TypeScript
This CL introduces a new 'WidgetElement' type. It is never
really constructed but it has the __widget and __widgetCounter
field. This way, TypeScript is able to understand
what is going on.

The rest of the CL are mostly added JSDoc annotation, casts and some
more explicit null checks that both TS and Closure understand.

CL 3 out of 3 to enable TS for ui/Widget.js.

Bug: chromium:1011811
Change-Id: I15f3918a25f3387d79ad06a68e79e427de5890e0