Display link in Application Panel for pages with no Service Workers
Currently, the Service Worker tab in the Application Panel in DevTools
does not show anything if the page does not have any service workers
registered for that origin. Hence, this CL fixes that by modifying the
CSS file for it. This is because the correct behaviour is that even
if there are no service workers for the page, it should still show
the link to chrome://serviceworker-internals to view all other
origins.

Bug: 1115410
Change-Id: I5723dcdddc793189f2fc0764801e5c212526a481