Format decoded X-Client-Data HTTP header values as code
This clarifies that this is the only content the header contains. See
the discussion here: https://crbug.com/1103854#c25

Screenshots:
- Before: https://i.imgur.com/LJHFIp9.png
- After: https://i.imgur.com/7mLKiEc.png

Note that these screenshots show this as a response header instead of a
request header, since that was easier to test. In real-world scenarios
it would be a request header instead, but the UI looks the same.

Bug: chromium:1103854
Change-Id: I06246ada6c61e8b8fc59c38760f79dc4357e0933