css: Remove our implementation of data-keyboard-focus
The web platform has :focus-visible now which does this for us.

https://developer.mozilla.org/en-US/docs/Web/CSS/:focus-visible

A previous CL changed all of our CSS selectors over to rely on
focus-visible instead.

For the use cases where we called markAsFocusedByKeyboard() manually
from UI components, these still work as intended because the
:focus-visible implementation tracks this for us so manually
focusing the element with .focus() will still result in :focus-visible.

The elementIsFocusedByKeyboard() calls were replaced by checking if
:focus-visible is on the element or not. This is not a very good way
to do this, but fixing it is outside the scope of this CL.

Change-Id: I429adf7b1e7d8a8f2520cc75d0e8a0e50cfada32