Add scroll to devtools for node connection screen.
This CL will fix a issue when the devtools for node is open and the
console is also open too high and the connection information is hidden.

This CL will add an scrollbar to show the users that there is more
content.

Before: https://i.imgur.com/MlRAyeA.gif
After: https://i.imgur.com/eABPcPM.gif

Bug: 1115655
Change-Id: I4e3507ada79bc38e5fe194c6aaea5cce2ca7252b