DevTools: Fix issues when reloading DevTools after setting change
Previously, when a setting change in devtools requires a reloading
and this devtools frontend is undocked. Clicking on the "Reload
DevTools" in the inspector view may cause 1) device mode button not
working and 2) a tab appears instead of that a frontend is reloaded.

The cause of the issue is that when "Reload DevTools" is clicked,
the reloading logic fails to consider the case when the devtools is
undocked.

Fix is to use the same logic as frontend.components.reload in
frontend.ui. We move the DockController.js from components to ui as
1) ui cannot depend on component and 2) DockController is arguably
a legitimate part of ui.

The after gif is below, where we tested both cases.
https://imgur.com/8LVLPMW

Bug: crbug/1105779
Bug: crbug/1113601
Change-Id: Ie8e1e60c49ff733de53ab7ad1409d8c44e99162a