DevTools: Fix issues when reloading DevTools after setting change
Devtools has some settings that requires a reloading of devtools
after changing them.

If devtools frontend is undocked, several issues were noticed.
1) device mode button does not work.
2) clicking on "Reload devtools" button, a tab appears instead of
that the devtools is reloaded.

The cause of the issue is that when "Reload DevTools" is clicked,
the reloading logic fails to consider the case when the devtools is
undocked.

Fix is to use the same logic as frontend.components.reload in
frontend.ui. We move the DockController.js from components to ui as
1) ui cannot depend on component and 2) DockController is arguably
a legitimate part of ui.

The after gif is below, where we tested both cases.
https://imgur.com/8LVLPMW

Bug: crbug/1105779
Bug: crbug/1113601
Change-Id: Ie8e1e60c49ff733de53ab7ad1409d8c44e99162a