Layout: Display class/id-less elements just with their tagname
When an element has no classes or ids, the Layout panel was still
previewing it as div. (with the dot at the end).
Now, with this change, we make sure that no empty classes are generated
in the UI.

Before: https://imgur.com/L0Gn7p5.jpg
After: https://imgur.com/nsdcUQF.jpg

Bug: 1116353
Change-Id: I2d8bb67f43bd003a3f80cd8c960c8a7d1e8d7992