Add getSurveyAPIKey to InspectorFrontendHost
This new method is used to get the API key to use for in-product
surveys. The key itself is in the chrome-internal repository and is
piped through in this chrome CL: http://crrev.com/c/2362182

We need to land this CL and wait for it to roll into chromium before
the chromium CL lands, as devtools_compatibility.js is used by the
backend.

Bug: 1112738
Change-Id: Ib44333bf7ca8ac226e7d4e31d6dd51f868144ab3