Reland "Fix path escaping"
This reverts commit a292e76f9ca0c1cd65fed699d552ea703e245a8a.
This relands commit 10892e33a22289e7a8c2006448962ce48e4d5867.

Inconsistent URL escaping was breaking snippets, which by default
have characters (space and #) which need to be escaped.

As a general rule, anything called 'url' or 'path' is now url-encoded,
and anything called 'name' or 'platformPath' is not.

Updated to match upper and lower case drive letters in Node on Windows.

Change-Id: Ib82bc34ad6769dfc9df6cef80c77401cee2a0783