Grid: Transform grid lines and labels based on the current writing mode
The backend now sends the writing-mode value for the highlighted grid.
See CL: https://chromium-review.googlesource.com/c/chromium/src/+/2363050/

By default, grid containers are horizontal-tb, unless the CSS specifies
otherwise. The only other 2 values that can be given by the backend are:
vertical-lr and vertical-rl.
Other values are either deprecated (tb, bt, ...) or not supported yet (sideways)

Based on the writing-mode, the overlay frontend applies a transformation to the
canvas context that draws the grid lines. It then saves this as a matrix so it
can be reused to transform the points where labels are placed (since those are
not part of the canvas).

Bug: 1096971
Change-Id: Ie7ec08850c51170fe973ed50d0787824792bdd16