Do not increase count in FrameManager if ResourceTreeModel is the same
This change ensures that the count in the FrameManager's map of frames
is only incremented when receiving a frameAdded event for a different
model. If receiving multiple FrameAdded events for the same model,
all but the first one are ignored.

The count could still become incorrect if events are received in
alternating order such as (model A, model B, model A). This could be
prevented by keeping track of all models a frame has been added to. For
now we trust the backend that such a scenario does not happen.

Change-Id: I1dc3677ae6649842785be0b6bb389dd98804b2fb