Component bridges presubmit check
This CL introduces a PRESUBMIT that will:

1. run `npm run regenerate-all-component-bridges`
2. Check that the git diff is clean.

If the git diff is not clean, it will fail. If the bridge has to be
manually edited (e.g. to work around a bug in the generator), it must
have a comment that matches the pattern `MANUALLY_EDITED_BRIDGE=...`.
The regeneration script will ignore bridges that contain these.

Change-Id: I372a3cd17a8e94bcdcea69ef64db189d2be6bde1