Query the model late for the target
The previous code was querying the model for the target at the creation
of the link to the elements panel. This defeated the purpose of the
getTargetIfNotDisposed method: That method protects against situations
where an issue is shown linking to an element in a frame who's target
was already detached (because, for example, it navigated OOP).

Bug: chromium:1115421
Change-Id: I1eb903e60a391e294a4b9ed5c99338b25ec77992