Sources: Restore breakpoints on formatted code
This CL fixes a regression introduced in Chrome 61(!) where breakpoints
in pretty-printed code appear to persist after reloading the DevTools,
but don't actually function. The issue was that the source formatter
added its UISourceCode containing the formatted code to the project
before storing the metadata that allows its source mapping to work, so
the breakpoints for the pretty-printed code would be set on the
formatted script rather than mapped to a location in the actual script.

Video of bug: https://i.imgur.com/OTDlq5g.mp4

Same thing with the fix: https://i.imgur.com/AZ7nT17.mp4

Bug: 930334
Change-Id: I53830839ed45a78d11ad2cc27fc5cc1bf09ec4a4