Add durations for individual presubmit steps
The time of a regular presubmit time has increased. To ease debugging
now and in the future, add the durations of each individual step
to the logs.

Since _checkWithNodeScript and all non-Node steps rely on
_ExecuteSubProcess we only need to implement it there.

Also reorder the presubmit outputs such that we only print "Running
X check" when we actually run it.

R=petermarshall@chromium.org

Bug: 1116989
Change-Id: I1f30c3b5dadb961ebfbfef56d50bd9347e9069f2