Reland "Localization: creating and integrating i18n module"
This reverts commit 8f012723b79deb820562b06e54bd34600578cef7.

The cause of the i18n not working on debug configuration was that
some files were missing from the grd lists.
- For i18nImpl I followed the same process as the other modules.

- For i18n-bundle.js(generated)
  an additional check is needed as the file does not exist as a source
  file.

To make it easy to review:
the first patchset contains the reland with no modifications the >2 patchset
contains the fixes, so if you diff Patchset 1 vs > 1 you should see the fixes
described above.

Change-Id: Iee3767b6afeb1a039be952f787886f94c0752f56