Grid: Support RTL column labels in the grid overlay
The grid overlay supports RTL directed grids, to some extent. The lines and gaps
are drawn correctly, but the number, names, sizing and area labels are
misplaced.

For labels to work, a backend change is required, this will be done in a
separate CL. But a couple of minor things also are required in the frontend:

- Making sure the logic to know if labels are too close to each other works with
  RTL (this requires comparing absolute distances).
- Making sure area name labels are within their areas (instead of starting at
  the right line, but going in the wrong direction, outside of the area).

These 2 things are fixed in this CL. And a new debug file is added for RTL.

Bug: 1096971
Change-Id: I5cbb21283f1eece2c377c4a5ad07e1850e3f36bf