Grid: Move the list of grids below settings in the Layout pane
There can be many grids on a page, which means a lot of scrolling is
required to access the settings in the Layout panel.
Also the settings section is really small, so moving it up top won't
make it considerably harder to access the grids.

Before: https://imgur.com/yfoFGmA.jpg
After: https://imgur.com/nOwhC1D.jpg

Bug: 1122493
Change-Id: I0d42527ca35de32468656ac440415dd226fc06e2