Fix blur handler of BreakPointEditDialog
When the users clicks on the `this.element` div, then the `blur`
handler of `this.contentElement` fires as it looses focus. Inside
the blur handler, we look at the element the will receive focus,
i.e. `relatedTarget`. The handler compares the `relatedTarget`
to `this.element`, not dismissing the dialog.

Since the `blur` handler is installed on this.contentElement, it
will no longer fire, even if `this.element` looses focus, causing
the dialog to stay open.

The solution is to install the `blur` handler this.element, to stay
consistent with the check inside the `blur` handler.

R=mathias@chromium.org

Bug: chromium:1081452
Change-Id: I84b01901110d0ce2acda088d3d8fc420afea6b7a