Fix URL flickering in image preview pane
https://chromium-review.googlesource.com/c/chromium/src/+/1633989 introduced
a regression where the alt text was visible for a short period of time,
until the actual image would be loaded.

This hapened, because the alt text was set in the constructor, whereas
the image itself is loaded in `_updateContentIfNeeded`. Since the browser
shows the alt text when there is no src (or the image hasn't finished loading
yet), this would flicker.

To remedy the situation, set the alt text after the src to make sure we do
have the image loaded before we also update the alt text.

R=petermarshall@chromium.org

Bug: chromium:1066415
Change-Id: I1391919560e503edddf6f7621a14d6d9a2281d95
Fixed: 1114617