Use `DOMMatrix` instead of `WebKitCSSMatrix` or `CSSMatrix`
`DOMMatrix` is the built-in, standard interface. `WebKitCSSMatrix`
is a legacy alias. `CSSMatrix` is a DevTools-internal alias that we
used for Closure typings, but since it doesn’t actually exist on the
Web Platform, these typings confuse the TypeScript compiler.

Instead of expecting everyone to figure this out like in [1], this
patch aligns on the standard `DOMMatrix` name, resulting in slightly
smoother TypeScriptification from here on out.

[1]: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2382857

Bug: chromium:1011811
Change-Id: I51cd6de03642136dd4041b6c9e6bd08c313e77db