Remove protocol.json from the root directory
This file appears to be old and now unused. I have tried searching for
it in both the Chromium codesearch and externally and no references
seem to linger around. Since the canonical location of CDP is
https://chromedevtools.github.io/devtools-protocol/, we should remove
our reference. (Although the canonical location is already the number
one search result and well-known)

R=yangguo@chromium.org

Change-Id: Ifc0241a63d2d2ad38966808df7264bc481b2eebe