Add warning about clearing data in Lighthouse panel
Lighthouse users are sometimes unaware that Lighthouse when clears
potentially important resources (IndexedDB, Web SQL, local storage).
This CL adds a warning section and warning message to inform the user
when important data may be cleared.

Additionally, there will be a warning prompt when any warnings are
present in case the user misses the warnings. This prompt can be ignored
through Lighthouse settings.

Bug: 806488
Change-Id: I64fd9e197868f690af0e0b89de52e0b7baa3f1fa