[ts] Add helpers to assert element type
This CL adds two helpers assertElement and assertElements that
can be used in conjunction with querySelector and querySelectorAll
to assert the element types of the results. The functions rely on
instanceof (thereby also performing a null check), and have types
that makes TypeScript reflect the check in the type system (i.e.
after asserting, e.g. assertElement(e, HTMLInputElement) TypeScript
knows that e is of type HTMLInputElement.

The CL also applies the function in a couple of places, and found
a bug in test/unittests/front_end/elements/LayoutPane_test.ts where
a HTMLButtonElement as incorrectly asserted as HTMLInputElement.

Bug: chromium:1011811
Change-Id: Icca56588a75597879003a298040e061fa088da78