Fix 404 layoutPane.css in unit tests
This CL does two things:

1. Makes the tests fail if a CSS file fails to preload correctly.
2. Fixes the 404 by adding layoutPane.css to the list of files to be copied.

It's not good that the copying of the CSS is done by the
unit_test/frontend/BUILD.gn. A follow up CL will move the copying of CSS
files into the respective BUILD.gns of the source code (e.g. `ui/*.css`
will be copied to gen in `ui/BUILD.gn`).

Change-Id: Ia46cbcd2b00757b62fce3b5b128a01c30df0281d