Grid: New histogram to track how many grids users highlight in the page
There is a backend part to this change:
https://chromium-review.googlesource.com/c/chromium/src/+/2379254/

The goal is to track how many grids people highlight in web pages using either
the adorner in the Elements tree, or the checkboxes in the Layout panel.

The histogram is done in a way where we track when users highlight 0, 1, 2, 3, 4
grids, and then between 5 and 9, 10 and 19, 20 and 49, and more than 50.
We have more granularity at the bottom end of the spectrum as the expectation is
that users will mostly highlight few grids at a time.

Bug: 1047356
Change-Id: If3a5282d3909e85ba73b96714fae1d4e2bc9ba36