[DevTools] Shorthand properties bug fix
If user creates a new property which overloads the expanded longhand property for a CSS property
defined in the same section, the longhand property is not getting overloaded in the styles pane.
Also if the user disables the overloaded property, the shorthand property is removed from the expanded
list. This should fix both the issues.
https://i.imgur.com/EDYbFzn.png
Bug:1002914

Change-Id: Ib633a3deb45b294a4504a4e8af4d1632d4f9a975